+++ testing.R
@@ -1,29 +1,11 @@
 devtools::load_all()
 
-# Load library and data
 library(rsimsum)
-data("MIsim2", package = "rsimsum")
-s <- simsum(data = MIsim2, estvarname = "b", true = 0.5, se = "se", methodvar = c("m1", "m2"))
-s
-summary(s, stats = "bias")
-autoplot(summary(s), type = "zip")
+s2 <- simsum(data = relhaz, estvarname = "theta", true = -0.50, se = "se", methodvar = "model", by = c("baseline", "n"))
+a <- print(summary(s2, stats = "bias"))
 
-
-data(frailty2)
-ms <- multisimsum(
-  data = frailty2, par = "par", true = c(
-    trt = -0.50,
-    fv = 0.75
-  ),
-  estvarname = "b",
-  se = "se",
-  methodvar = c("m_baseline", "m_frailty"),
-  by = "fv_dist"
-)
-ms
-summary(ms)
-
-
+library(knitr)
+kable(a[[1]], caption = names(a)[1], align = "r")
 
 ### All-in-one testing
 devtools::document()